chore(ss-management): Move cypress configs to cypress folder (similar to https://github.com/elastic/kibana/pull/165094)
