+++ x-pack/plugins/security_solution/public/management/cypress/cypress_endpoint.config.ts
@@ -6,10 +6,8 @@
  */
 
 import { defineCypressConfig } from '@kbn/cypress-config';
-// eslint-disable-next-line @kbn/imports/no_boundary_crossing
-import { dataLoaders, dataLoadersForRealEndpoints } from './cypress/support/data_loaders';
-// eslint-disable-next-line @kbn/imports/no_boundary_crossing
-import { responseActionTasks } from './cypress/support/response_actions';
+import { dataLoaders, dataLoadersForRealEndpoints } from './support/data_loaders';
+import { responseActionTasks } from './support/response_actions';
 
 // eslint-disable-next-line import/no-default-export
 export default defineCypressConfig({
