+++ x-pack/plugins/security_solution/public/management/cypress/tsconfig.json
@@ -2,9 +2,6 @@
   "extends": "../../../../../../tsconfig.base.json",
   "include": [
     "**/*",
-    "../cypress_endpoint.config.ts",
-    "../cypress.config.ts",
-    "./cypress.d.ts"
   ],
   "exclude": [
     "target/**/*"
