+++ x-pack/plugins/security_solution/tsconfig.json
@@ -16,9 +16,7 @@
   ],
   "exclude": [
     "target/**/*",
-    "**/cypress/**",
-    "public/management/cypress_endpoint.config.ts",
-    "public/management/cypress.config.ts"
+    "**/cypress/**"
   ],
   "kbn_references": [
     "@kbn/core",
